fix(landing): extract cached GitHub data for lazy-loaded open-source
Move CachedGithubData with use cache into a separate server module so
next/dynamic lazy loading of open-source does not violate Client Component
constraints.
